+++ chain/actors/builtin/market/v0.go
@@ -302,7 +302,7 @@ func (r *publishStorageDealsReturn0) DealIDs() ([]abi.DealID, error) {
 	return r.IDs, nil
 }
 
-func (s *state0) GetAllocationId(dealId abi.DealID) (verifregtypes.AllocationId, error) {
+func (s *state0) GetAllocationIdForPendingDeal(dealId abi.DealID) (verifregtypes.AllocationId, error) {
 
 	return 0, xerrors.Errorf("unsupported before actors v9")
