+++ chain/actors/builtin/market/v3.go
@@ -297,7 +297,7 @@ func (r *publishStorageDealsReturn3) DealIDs() ([]abi.DealID, error) {
 	return r.IDs, nil
 }
 
-func (s *state3) GetAllocationId(dealId abi.DealID) (verifregtypes.AllocationId, error) {
+func (s *state3) GetAllocationIdForPendingDeal(dealId abi.DealID) (verifregtypes.AllocationId, error) {
 
 	return 0, xerrors.Errorf("unsupported before actors v9")
