+++ chain/actors/builtin/market/v5.go
@@ -297,7 +297,7 @@ func (r *publishStorageDealsReturn5) DealIDs() ([]abi.DealID, error) {
 	return r.IDs, nil
 }
 
-func (s *state5) GetAllocationId(dealId abi.DealID) (verifregtypes.AllocationId, error) {
+func (s *state5) GetAllocationIdForPendingDeal(dealId abi.DealID) (verifregtypes.AllocationId, error) {
 
 	return 0, xerrors.Errorf("unsupported before actors v9")
