+++ chain/actors/builtin/market/v6.go
@@ -315,7 +315,7 @@ func (r *publishStorageDealsReturn6) DealIDs() ([]abi.DealID, error) {
 	return r.IDs, nil
 }
 
-func (s *state6) GetAllocationId(dealId abi.DealID) (verifregtypes.AllocationId, error) {
+func (s *state6) GetAllocationIdForPendingDeal(dealId abi.DealID) (verifregtypes.AllocationId, error) {
 
 	return 0, xerrors.Errorf("unsupported before actors v9")
