+++ chain/actors/builtin/market/v8.go
@@ -332,7 +332,7 @@ func (r *publishStorageDealsReturn8) DealIDs() ([]abi.DealID, error) {
 	return r.IDs, nil
 }
 
-func (s *state8) GetAllocationId(dealId abi.DealID) (verifregtypes.AllocationId, error) {
+func (s *state8) GetAllocationIdForPendingDeal(dealId abi.DealID) (verifregtypes.AllocationId, error) {
 
 	return 0, xerrors.Errorf("unsupported before actors v9")
