+++ chain/actors/builtin/verifreg/v2.go
@@ -90,7 +90,7 @@ func (s *state2) GetState() interface{} {
 	return &s.State
 }
 
-func (s *state2) GetAllocation(addr address.Address, allocationId verifreg9.AllocationId) (*verifreg9.Allocation, bool, error) {
+func (s *state2) GetAllocation(clientIdAddr address.Address, allocationId verifreg9.AllocationId) (*verifreg9.Allocation, bool, error) {
 
 	return nil, false, xerrors.Errorf("unsupported in actors v2")
