+++ chain/actors/builtin/verifreg/v4.go
@@ -91,7 +91,7 @@ func (s *state4) GetState() interface{} {
 	return &s.State
 }
 
-func (s *state4) GetAllocation(addr address.Address, allocationId verifreg9.AllocationId) (*verifreg9.Allocation, bool, error) {
+func (s *state4) GetAllocation(clientIdAddr address.Address, allocationId verifreg9.AllocationId) (*verifreg9.Allocation, bool, error) {
 
 	return nil, false, xerrors.Errorf("unsupported in actors v4")
