+++ chain/actors/builtin/verifreg/v5.go
@@ -91,7 +91,7 @@ func (s *state5) GetState() interface{} {
 	return &s.State
 }
 
-func (s *state5) GetAllocation(addr address.Address, allocationId verifreg9.AllocationId) (*verifreg9.Allocation, bool, error) {
+func (s *state5) GetAllocation(clientIdAddr address.Address, allocationId verifreg9.AllocationId) (*verifreg9.Allocation, bool, error) {
 
 	return nil, false, xerrors.Errorf("unsupported in actors v5")
