+++ documentation/en/api-v1-unstable-methods.md
@@ -179,8 +179,8 @@
   * [StateEncodeParams](#StateEncodeParams)
   * [StateGetActor](#StateGetActor)
   * [StateGetAllocation](#StateGetAllocation)
+  * [StateGetAllocationForPendingDeal](#StateGetAllocationForPendingDeal)
   * [StateGetBeaconEntry](#StateGetBeaconEntry)
-  * [StateGetDealAllocation](#StateGetDealAllocation)
   * [StateGetNetworkParams](#StateGetNetworkParams)
   * [StateGetRandomnessFromBeacon](#StateGetRandomnessFromBeacon)
   * [StateGetRandomnessFromTickets](#StateGetRandomnessFromTickets)
@@ -5728,7 +5728,7 @@ Response:
 ```
 
 ### StateGetAllocation
-StateGetAllocation returns the allocation ID for a given address and allocation ID.
+StateGetAllocation returns the allocation for a given address and allocation ID.
 
 
 Perms: read
@@ -5738,6 +5738,42 @@ Inputs:
 [
   "f01234",
   0,
+  [
+    {
+      "/": "bafy2bzacea3wsdh6y3a36tb3skempjoxqpuyompjbmfeyf34fi3uy6uue42v4"
+    },
+    {
+      "/": "bafy2bzacebp3shtrn43k7g3unredz7fxn4gj533d3o43tqn2p2ipxxhrvchve"
+    }
+  ]
+]
+```
+
+Response:
+```json
+{
+  "Client": 1000,
+  "Provider": 1000,
+  "Data": {
+    "/": "bafy2bzacea3wsdh6y3a36tb3skempjoxqpuyompjbmfeyf34fi3uy6uue42v4"
+  },
+  "Size": 1032,
+  "TermMin": 10101,
+  "TermMax": 10101,
+  "Expiration": 10101
+}
+```
+
+### StateGetAllocationForPendingDeal
+StateGetAllocationForPendingDeal returns the allocation for a given deal ID of a pending deal.
+
+
+Perms: read
+
+Inputs:
+```json
+[
+  5432,
   [
     {
       "/": "bafy2bzacea3wsdh6y3a36tb3skempjoxqpuyompjbmfeyf34fi3uy6uue42v4"
@@ -5784,42 +5820,6 @@ Response:
 {
   "Round": 42,
   "Data": "Ynl0ZSBhcnJheQ=="
-}
-```
-
-### StateGetDealAllocation
-StateGetDealAllocation returns the allocation ID for a given deal ID.
-
-
-Perms: read
-
-Inputs:
-```json
-[
-  5432,
-  [
-    {
-      "/": "bafy2bzacea3wsdh6y3a36tb3skempjoxqpuyompjbmfeyf34fi3uy6uue42v4"
-    },
-    {
-      "/": "bafy2bzacebp3shtrn43k7g3unredz7fxn4gj533d3o43tqn2p2ipxxhrvchve"
-    }
-  ]
-]
-```
-
-Response:
-```json
-{
-  "Client": 1000,
-  "Provider": 1000,
-  "Data": {
-    "/": "bafy2bzacea3wsdh6y3a36tb3skempjoxqpuyompjbmfeyf34fi3uy6uue42v4"
-  },
-  "Size": 1032,
-  "TermMin": 10101,
-  "TermMax": 10101,
-  "Expiration": 10101
 }
 ```
