+++ itests/verifreg_test.go
@@ -123,10 +123,7 @@ func TestVerifiedClientTopUp(t *testing.T) {
 			//stm: @CHAIN_STATE_VERIFIED_CLIENT_STATUS_001
 			dcap, err := api.StateVerifiedClientStatus(ctx, verifiedClientAddr, types.EmptyTSK)
 			require.NoError(t, err)
-
-			if !dcap.Equals(datacap) {
-				t.Fatal("")
-			}
+			require.Equal(t, *dcap, datacap)
 
 			// try to assign datacap to the same client should fail for actor v4 and below
 			params, err = actors.SerializeParams(&verifregst.AddVerifiedClientParams{Address: verifiedClientAddr, Allowance: datacap})
@@ -280,10 +277,7 @@ func TestRemoveDataCap(t *testing.T) {
 	//stm: @CHAIN_STATE_VERIFIED_CLIENT_STATUS_001
 	dcap, err := api.StateVerifiedClientStatus(ctx, verifiedClientAddr, types.EmptyTSK)
 	require.NoError(t, err)
-
-	if !dcap.Equals(datacap) {
-		t.Fatal("unexpected datacap")
-	}
+	require.Equal(t, *dcap, datacap)
 
 	// helper to create removedatacap message
 	makeRemoveDatacapMsg := func(removeDatacap big.Int, proposalID uint64) *types.Message {
@@ -354,10 +348,7 @@ func TestRemoveDataCap(t *testing.T) {
 	//stm: @CHAIN_STATE_VERIFIED_CLIENT_STATUS_001
 	dcap, err = api.StateVerifiedClientStatus(ctx, verifiedClientAddr, types.EmptyTSK)
 	require.NoError(t, err)
-
-	if !dcap.Equals(big.Sub(datacap, removeDatacap)) {
-		t.Fatal("unexpected datacap")
-	}
+	require.Equal(t, *dcap, big.Sub(datacap, removeDatacap))
 
 	// now take away the second half!
